requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
sentence-transformers
transformers
faker
pydantic
python-dotenv
psycopg2-binary
